requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.13
django-extensions==2.2.9
django-rest-swagger==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
idna==2.9
importlib-metadata==1.6.1
inflection==0.5.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
model-bakery==1.1.1
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
openapi-codec==1.3.2
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.4
webencodings==0.5.1
xmltodict==0.12.0
zipp==3.1.0
